fix tail heights on click
Refactored draw lines login into drawLines(), so that this could be called during click after hiding or showing.

This fixes tail heights that get drawn too high due to collapsed siblings. Also fixes heights being set to 0 in case $('table').treetable() is called on a table that is currently hidden.

I realize this is not the most efficient fix, since only certain lines need to be re-adjusted depending upon the parent/sibling context.